Updates to a Flask Example
