+++ Python/Run_Flask.bat
@@ -1,0 +1,4 @@
+@ECHO OFF
+SET FLASK_APP=FlaskJSONExample.py
+
+flask run
